게시글 구현 수정

+++ src/main/java/com/stock/yu/downbitbe/board/application/PostService.java
@@ -18,40 +18,42 @@ public class PostService {
     private final PostRepository postRepository;
 
     @Transactional(readOnly = true)
-    public PostDto findPostByPostId(Long postId){
+    public PostResponseDto findPostByPostId(Long postId){
         Post post = postRepository.findById(postId).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
-        return new PostDto(post);
+        return new PostResponseDto(post);
     }
 
     @Transactional(readOnly = true)
-    public List<PostListDto> findAllPostsById(Long boardId) {
+    public List<PostListResponseDto> findAllPostsById(Long boardId) {
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new IllegalArgumentException("게시판이 존재하지 않습니다."));
         return postRepository.findAllByBoardId(boardId).stream()
-                .map(PostListDto::new).collect(Collectors.toList());
+                .map(PostListResponseDto::new).collect(Collectors.toList());
     }
 
     @Transactional
-    public Long createPost(PostCreateDto postCreateDto, Long boardId) {
+    public Long createPost(PostCreateRequestDto postCreateRequestDto, Long boardId, User user) {
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new IllegalArgumentException("게시판이 존재하지 않습니다."));
-        User user = new User(); // 토큰 사용 후 사용자 정보 받기
-        Post post = postCreateDto.toEntity(board, user);
+
+        Post post = postCreateRequestDto.toEntity(board, user);
         return postRepository.save(post).getId();
     }
 
-//    @Transactional
-//    public Long updatePost(PostDto postDto, Long postId) {
-//        Post post = postRepository.findById(postId).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
-//        User user = new User();     // 사용자 정보 확인 후 작성자만 수정 가능
-//
-//    }
+    @Transactional
+    public Long updatePost(PostUpdateRequestDto postUpdateRequestDto, Long postId, User user) {
+        Post post = postRepository.findById(postId).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
+        if(!post.getUser().getUserId().equals(user.getUserId())){
+            throw new RuntimeException("작성자와 일치하지 않습니다.");
+        }
+        return postRepository.save(post.updatePost(postUpdateRequestDto.toEntity())).getId();
+    }
 
     @Transactional
-    public Long deletePost(Long postId){
+    public Long deletePost(Long postId, User user){
         Post post = postRepository.findById(postId).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
-        User user = new User(); // 사용자 정보 확인 후 작성자 or admin 삭제 가능
+        if(!post.getUser().getUserId().equals(user.getUserId())){
+            throw new RuntimeException("작성자와 일치하지 않습니다.");
+        }
         postRepository.delete(post);
         return post.getId();
     }
-
-
 }
